Fixed logout button in settings page

+++ public/assets/js/connection.js
@@ -2,19 +2,19 @@ $('document').ready(function () {
   // Initiates when connectBtn is clicked
   $(".connectBtn").on("click", function () {
     var id = $(this).data("id") // Grabs the other user's id
-    var connectReq = $(this).data("connection") // Boolean
+    var connected = $(this).data("connection") // Boolean
 
-    // If user is trying to connect make a PUT request
-    if (connectReq) {
-      $.ajax("/connect", {
-        type: "PUT",
+    // If user is already connected make a DELETE request
+    if (connected) {
+      $.ajax("/disconnect", {
+        type: "DELETE",
         data: { associateId: id }
       }).then(() => {
         location.reload();
       });
-    } else { // Otherwise make a DELETE request
-      $.ajax("/disconnect", {
-        type: "DELETE",
+    } else { // Otherwise make a PUTrequest
+      $.ajax("/connect", {
+        type: "PUT",
         data: { associateId: id }
       }).then(() => {
         location.reload();
